requirements.txt
boto3==1.38.14
botocore==1.38.14
python-dotenv==1.1.0
selenium==4.32.0